Fix 32-bit Raspbian build failures
`LLVMSizeOfTypeInBits` returns `unsigned long long`.  Assigning it to
`size_t` narrows on 32-bit targets where `size_t` is 32 bits, which
`-Werror=conversion` rejects.  Use the actual return type instead.

`-Wno-unknown-warning-option` is Clang-only.  GCC doesn't recognise it,
and `-Werror` promotes the unrecognised flag to a fatal error.  Guard it
behind a Clang compiler-id check so the GCC build path documented in
BUILD.md works.

+++ src/libponyc/codegen/gentagged.c
@@ -32,7 +32,8 @@ bool gentagged_is_taggable(compile_t* c, reach_type_t* t)
   // Any numeric primitive fitting in 32 bits is taggable.  On LP64 (Linux,
   // macOS) that covers Bool, U8–U32, I8–I32, and F32.  On LLP64 (Windows)
   // ILong and ULong are also 32-bit and therefore taggable.
-  size_t prim_bits = LLVMSizeOfTypeInBits(c->target_data, c_t->primitive);
+  unsigned long long prim_bits =
+    LLVMSizeOfTypeInBits(c->target_data, c_t->primitive);
   return prim_bits <= 32;
 }
